rename IdGetter to IdRangeGetter. rename DatabaseIdGetter to DatabaseIdRangeGetter.
